+++ src/main/java/me/myogoo/ae2fct/init/AE2FCTItems.java
@@ -25,7 +25,7 @@ public final class AE2FCTItems {
 
     private static <T extends Item> RegistryObject<T> registerItem(String name, Supplier<T> supplier) {
         var item = ITEMS.register(name, supplier);
-        MyotusAPI.creativeTabRegistrar().creativeTabItem(item);
+        MyotusAPI.creativeTabs().registerCreativeTabItem(item);
         return item;
     }
 }